+++ src/components/charts-generic/constants.ts
@@ -1,4 +1,4 @@
-export const LEFT_MARGIN_OFFSET = 5;
+export const LEFT_MARGIN_OFFSET = 10;
 export const BOTTOM_MARGIN_OFFSET = 10;
 export const MINI_CHART_WIDTH = 350;